backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pymongo==4.9.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.7
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
dnspython==2.7.0